Fourth version uploaded works
Ten random numbers were generated between 0 and 100000 and the output predicted. All ten of these tests passed with upper and lower indices of 9 and 4 respectively. Then a couple other tests were conducted with previous numbers but different index boundaries. Specifically, the file was executed with the instruction 'revbits4 61680 11 4' and the output was '65280'. Additionally, the program was designed to take exactly 3 command line arguments (other than the program name itself).